requirements.txt
Django>=5.2,<5.3
python-dotenv>=1.0.0
boto3>=1.34.0
botocore>=1.34.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0 